move files to consulting&sponsor
