+++ controllers/Consulting&SponsorsControllers/consultingController.js
@@ -1,4 +1,4 @@
-const consulting = require("../models/consultingMd");
+const consulting = require("../../models/consultingMd");
 
 const consultingPost = (req, res, next) => {
   const issue = req.body.issue;
